feat: add mock classes to base
